+++ client/src/app/+my-account/my-account-settings/my-account-profile/my-account-profile.component.ts
@@ -4,7 +4,7 @@ import { FormReactive, UserService } from '../../../shared'
 import { User } from '@app/shared'
 import { I18n } from '@ngx-translate/i18n-polyfill'
 import { FormValidatorService } from '@app/shared/forms/form-validators/form-validator.service'
-import { Subject } from 'rxjs/Subject'
+import { Subject } from 'rxjs'
 import { UserValidatorsService } from '@app/shared/forms/form-validators/user-validators.service'
 
 @Component({