+++ client/src/app/videos/+video-edit/shared/video-edit.component.ts
@@ -35,7 +35,6 @@ export class VideoEditComponent implements OnInit {
 
   schedulePublicationEnabled = false
 
-  error: string = null
   calendarLocale: any = {}
   minScheduledDate = new Date()