+++ client/src/assets/player/utils.ts
@@ -1,4 +1,3 @@
-import { is18nLocale, isDefaultLocale } from '../../../../shared/models/i18n/i18n'
 import { VideoFile } from '../../../../shared/models/videos'
 
 function toTitleCase (str: string) {
@@ -18,60 +17,6 @@ function bytes (value) {
   const calc = Math.floor(value / (format.max / 1024)).toString()
 
   return [ calc, format.type ]
-}
-
-function getStoredVolume () {
-  const value = getLocalStorage('volume')
-  if (value !== null && value !== undefined) {
-    const valueNumber = parseFloat(value)
-    if (isNaN(valueNumber)) return undefined
-
-    return valueNumber
-  }
-
-  return undefined
-}
-
-function getStoredMute () {
-  const value = getLocalStorage('mute')
-  if (value !== null && value !== undefined) return value === 'true'
-
-  return undefined
-}
-
-function getAverageBandwidth () {
-  const value = getLocalStorage('average-bandwidth')
-  if (value !== null && value !== undefined) {
-    const valueNumber = parseInt(value, 10)
-    if (isNaN(valueNumber)) return undefined
-
-    return valueNumber
-  }
-
-  return undefined
-}
-
-function getStoredTheater () {
-  const value = getLocalStorage('theater-enabled')
-  if (value !== null && value !== undefined) return value === 'true'
-
-  return undefined
-}
-
-function saveVolumeInStore (value: number) {
-  return setLocalStorage('volume', value.toString())
-}
-
-function saveMuteInStore (value: boolean) {
-  return setLocalStorage('mute', value.toString())
-}
-
-function saveTheaterInStore (enabled: boolean) {
-  return setLocalStorage('theater-enabled', enabled.toString())
-}
-
-function saveAverageBandwidth (value: number) {
-  return setLocalStorage('average-bandwidth', value.toString())
 }
 
 function isMobile () {
@@ -132,39 +77,15 @@ function videoFileMinByResolution (files: VideoFile[]) {
   return min
 }
 
+// ---------------------------------------------------------------------------
+
 export {
   toTitleCase,
   buildVideoLink,
-  getStoredVolume,
-  saveVolumeInStore,
-  saveAverageBandwidth,
-  getAverageBandwidth,
-  saveMuteInStore,
   buildVideoEmbed,
-  getStoredMute,
   videoFileMaxByResolution,
   videoFileMinByResolution,
   copyToClipboard,
-  getStoredTheater,
-  saveTheaterInStore,
   isMobile,
   bytes
-}
-
-// ---------------------------------------------------------------------------
-
-const KEY_PREFIX = 'peertube-videojs-'
-
-function getLocalStorage (key: string) {
-  try {
-    return localStorage.getItem(KEY_PREFIX + key)
-  } catch {
-    return undefined
-  }
-}
-
-function setLocalStorage (key: string, value: string) {
-  try {
-    localStorage.setItem(KEY_PREFIX + key, value)
-  } catch { /* empty */ }
 }